[New Feature] Comment and Meme deletion by Admin
Coded a new feature that allows administrator users to delete memes and
comments. I have used AJAX requests at the Front-end and modfied the
appropriate controller's methods to return a JSON encoded response of the
result. I have put a mechanism in place, that verifies the user's roles,
and based on that, renders the views and handles the AJAX requests at the
endpoints.